refactor: remove id field in PageIndexInfo
