+++ packages/core/src/node/route/RoutePage.ts
@@ -1,0 +1,15 @@
+import type { RouteMeta } from '@rspress/shared';
+
+export class RoutePage {
+  routeMeta: RouteMeta;
+  // TODO: add pageIndexInfo
+  // pageIndexInfo: PageIndexInfo;
+
+  static create(routeMeta: RouteMeta) {
+    return new RoutePage(routeMeta);
+  }
+
+  constructor(routeMeta: RouteMeta) {
+    this.routeMeta = routeMeta;
+  }
+}
