+++ packages/core/src/node/route/extractPageData.test.ts
@@ -1,9 +1,9 @@
 import { join } from 'node:path';
 import { describe, expect, it } from 'vitest';
-import type { RouteService } from '../../route/RouteService';
+import type { RouteService } from './RouteService';
 import { extractPageData } from './extractPageData';
 
-const fixtureBasicDir = join(__dirname, '../../route/fixtures/basic');
+const fixtureBasicDir = join(__dirname, './fixtures/basic');
 
 const absolutize = (relativePath: string) => {
   return join(fixtureBasicDir, `.${relativePath}`);
@@ -12,10 +12,6 @@ const absolutize = (relativePath: string) => {
 describe('extractPageData', async () => {
   it('basic', async () => {
     const pageData = await extractPageData(
-      [],
-      {},
-      'http://localhost:3000',
-      fixtureBasicDir,
       {
         getRoutes: () =>
           Object.values({
@@ -53,7 +49,13 @@ describe('extractPageData', async () => {
             },
           }),
       } as RouteService,
-      false,
+      {
+        alias: {},
+        domain: 'http://localhost:3000',
+        replaceRules: [],
+        root: fixtureBasicDir,
+        searchCodeBlocks: false,
+      },
     );
     expect(pageData).toMatchInlineSnapshot(`
       [
@@ -61,14 +63,12 @@ describe('extractPageData', async () => {
           "_filepath": "<ROOT>/packages/core/src/node/route/fixtures/basic/a.mdx",
           "_flattenContent": "# Page a
       ",
-          "_html": "<h1 id="page-a">Page a<a aria-hidden="true" href="#page-a">#</a></h1>",
           "_relativePath": "a.mdx",
           "content": "#",
           "domain": "http://localhost:3000",
           "frontmatter": {
             "__content": undefined,
           },
-          "id": 0,
           "lang": "",
           "routePath": "/a",
           "title": "Page a",
@@ -79,14 +79,12 @@ describe('extractPageData', async () => {
           "_filepath": "<ROOT>/packages/core/src/node/route/fixtures/basic/guide/b.mdx",
           "_flattenContent": "# Page b
       ",
-          "_html": "<h1 id="page-b">Page b<a aria-hidden="true" href="#page-b">#</a></h1>",
           "_relativePath": "guide/b.mdx",
           "content": "#",
           "domain": "http://localhost:3000",
           "frontmatter": {
             "__content": undefined,
           },
-          "id": 1,
           "lang": "",
           "routePath": "/guide/b",
           "title": "Page b",
@@ -96,12 +94,10 @@ describe('extractPageData', async () => {
         {
           "_filepath": "<ROOT>/packages/core/src/node/route/fixtures/basic/guide/c.tsx",
           "_flattenContent": "",
-          "_html": "",
           "_relativePath": "guide/c.tsx",
           "content": "",
           "domain": "http://localhost:3000",
           "frontmatter": {},
-          "id": 2,
           "lang": "",
           "routePath": "/guide/c",
           "title": "",
@@ -112,14 +108,12 @@ describe('extractPageData', async () => {
           "_filepath": "<ROOT>/packages/core/src/node/route/fixtures/basic/index.mdx",
           "_flattenContent": "# homePage
       ",
-          "_html": "<h1 id="homepage">homePage<a aria-hidden="true" href="#homepage">#</a></h1>",
           "_relativePath": "index.mdx",
           "content": "#",
           "domain": "http://localhost:3000",
           "frontmatter": {
             "__content": undefined,
           },
-          "id": 3,
           "lang": "",
           "routePath": "/",
           "title": "homePage",
